Add prototype frontend w/ simple html
Added React components to load along with the Spring connection
test. It can upload recipes to a mongodb database after being
handled by the Spring framework. I believe it to be a satisfactory
test of all 3 systems. Note that this code is terrible and should
not come anywhere near the actual production code.